This just messed it up
That did not work, now the sample dot that should be at 0,0 is at -1k on the horizontal and 0 on the vertical.
In addition, the horizontal lines are shifted 1/2 grid to the left...
Meaning they stick out to the left by 1/2 grid length and there is a 1/2 grid length gap between the horizontal lines and the right edge.